iCub_simulator_tools/python_scripts/requirements.txt
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
EasyProcess==0.3
entrypoint2==0.2.3
entrypoints==0.3
h5py==3.1.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
jeepney==0.4.3
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.0
mistune==0.8.4
mss==6.0.0
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
opencv-python==4.4.0.44
packaging==20.3
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
progress==1.5
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
pyscreenshot==2.2
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
qtconsole==4.7.4
QtPy==1.9.0
Send2Trash==1.5.0
sentinel==0.1.2
six==1.14.0
tabulate==0.8.7
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
yarp==0.4.0
